+++ fixtures/proofs.php
@@ -84,7 +84,7 @@ return static function($prove) {
     yield $prove
         ->properties(
             'Counter properties',
-            Set\Properties::any(
+            Set::properties(
                 DownAndUpIsAnIdentityFunction::any(),
                 DownChangeState::any(),
                 LowerBoundAtZero::any(),